refactor: finish moving from post-processing renderer to render pipeline
