+++ systems/renderer/index.js
@@ -6,4 +6,3 @@ export { default as helper } from "./helper.js";
 export { default as line } from "./line.js";
 export { default as standard } from "./standard.js";
 export { default as skybox } from "./skybox.js";
-export { default as postProcessing } from "../render-pipeline/post-processing-renderer.js";
